Rename sample_utilities.py to utilities.py
