+++ sdk/containerregistry/azure-containerregistry/samples/async_samples/sample_list_tags_async.py
@@ -34,7 +34,7 @@ import asyncio
 import os
 from dotenv import find_dotenv, load_dotenv
 from azure.containerregistry.aio import ContainerRegistryClient
-from samples.sample_utilities import load_registry, get_authority, get_audience, get_credential
+from samples.utilities import load_registry, get_authority, get_audience, get_credential
 
 
 class ListTagsAsync(object):
