+++ sdk/containerregistry/azure-containerregistry/samples/sample_hello_world.py
@@ -31,7 +31,7 @@ USAGE:
 import os
 from dotenv import find_dotenv, load_dotenv
 from azure.containerregistry import ContainerRegistryClient
-from sample_utilities import load_registry, get_authority, get_audience, get_credential
+from utilities import load_registry, get_authority, get_audience, get_credential
 
 
 class HelloWorld(object):
