+++ sdk/containerregistry/azure-containerregistry/samples/sample_upload_download_manifest.py
@@ -28,7 +28,7 @@ from io import BytesIO
 from dotenv import find_dotenv, load_dotenv
 from azure.containerregistry import ContainerRegistryClient
 from azure.containerregistry._generated.models import Annotations, Descriptor, OCIManifest
-from sample_utilities import load_registry, get_authority, get_audience, get_credential
+from utilities import load_registry, get_authority, get_audience, get_credential
 
 
 class UploadDownloadManifest(object):
